Add BigCategoryList and MediumCategoryList and divide them into income and expense
